pyproject.toml
[project]
name = "dawson-court-scraper"
version = "3.0.0"
description = "US Tax Court opinion scraper with parallel document downloads"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.9.0",
    "aiofiles>=23.0.0",
    "tqdm>=4.66.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "tenacity>=8.2.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
dawson = "run:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "dawson_scraper/src"}

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "UP", "YTT", "B", "A", "C4", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true